e2e: remove tests 5 and 6
