--- e2e/test_007.py
@@ -1,70 +1,0 @@
-import random
-import os
-from threading import Thread, Lock
-from queue import Queue
-
-from mininet.log import info
-from minindn.minindn import Minindn
-from minindn.apps.app_manager import AppManager
-
-from fw import NDNd_FW
-import dv_util
-
-def scenario(ndn: Minindn, network='/minindn'):
-    """
-    Simple file transfer scenario with NDNd forwarder.
-    This tests routing convergence and cat/put operations.
-    """
-
-    info('Starting forwarder on nodes\n')
-    AppManager(ndn, ndn.net.hosts, NDNd_FW, network=network)
-
-    dv_util.setup(ndn, network=network)
-    dv_util.converge(ndn.net.hosts, network=network)
-
-    sample_size = min(8, len(ndn.net.hosts) - 1)
-    publish_size = min(2, sample_size - 1)
-    chatters = random.sample(ndn.net.hosts, sample_size)
-    chatter_names = {f"{network}/{node.name}" for node in chatters}
-
-    # set the svs forwarding strategy to replicast as the alo-latest example says to do
-    for node in ndn.net.hosts:
-        node.cmd("ndnd fw strategy-set prefix=/ndn/svs strategy=/localhost/nfd/strategy/multicast/v=1")
-
-
-    # start the alo client on each node
-    for node in chatters:
-        node.sendCmd(f"alo-latest {network}/{node.name}")
-
-    def receive_chat_message(node):
-        buf = []
-        while 1:
-            ch = node.read(1)
-            buf.append(ch)
-            if ch == "\n":
-                line = "".join(buf)
-                if line.split(":")[0] in chatter_names and "Skipping" not in line:
-                    return line
-                buf = []
-
-    try:
-        for node in chatters[:publish_size]:
-            # send a random message into chat
-            msg = os.urandom(8).hex()
-            node.write(msg + "\n")
-            info(f"{node.name} publishing {msg}\n")
-
-            for other_node in chatters:
-                if other_node.name == node.name:
-                    continue
-                received = receive_chat_message(other_node)
-                assert msg in received, "received wrong message"
-
-    # custom cleanup logic for this test
-    finally:
-        for node in chatters:
-            info(f"stopping alo-latest on {node.name}\n")
-            # send interrupt to alo-latest and reset the cmd stdout buffer read logic
-            node.popen(["pkill", "alo-latest"], stdout=-1).stdout.readline()
-            node.waiting = False
-            node.cmd("ls")
